first attempt to beautification
